requirements.txt
python-dotenv
flask
requests
```
alpha_vantage
Flask-SQLAlchemy PyMySQL
psycopg2-binary
Flask-Migrate
pydantic
```
bcrypt
#smtplib